BACKPORT: ASoC: Intel: avs: Fix build error on arc, m68k and sparc
On some platforms (i.e. arc, m68k and sparc) __fls returns an int (while
on most platforms it returns an unsigned long). This triggers a format
warning on these few platforms as the driver uses %ld to print a warning.

So explicitly cast the return value to unsigned long to make the warning
go away (and so fix allmodconfig build on the affected architectures).

Fixes: beed983621fb ("ASoC: Intel: avs: Machine board registration")
Link: https://lore.kernel.org/r/20220529141250.1979827-1-u.kleine-koenig@pengutronix.de

(cherry picked from commit d6f8168e4345cb00156743c19dc1a409126936e1)

Conflicts:
   sound/soc/intel/avs/board_selection.c

BUG=219931843
TEST=AVS works on APL and AML

+++ sound/soc/intel/avs/board_selection.c
@@ -288,7 +288,7 @@ static int avs_register_i2s_board(struct avs_dev *adev, struct snd_soc_acpi_mach
 	num_ssps = adev->hw_cfg.i2s_caps.ctrl_count;
 	if (fls(mach->link_mask) > num_ssps) {
 		dev_err(adev->dev, "Platform supports %d SSPs but board %s requires SSP%ld\n",
-			num_ssps, mach->drv_name, __fls(mach->link_mask));
+			num_ssps, mach->drv_name, (unsigned long)__fls(mach->link_mask));
 		return -ENODEV;
 	}
